Abstract out detecting DSA group memberships; update for supplementaryGid
gidNumber is no longer 800 for DDs. Instead we need to look at the list
of supplentaryGid and match on the strings "Debian" or "guest". This is
also a good time to abstract this logic out rather than doing
/if gidNumber==800/ all over the place.

For details of the field change, see https://lists.debian.org/debian-project/2015/10/msg00007.html

Signed-off-by: Jonathan Wiltshire <jmw@debian.org>