Remove "AM shouldn't know applicant" requirement
It's true that everyone who knows the applicant and thinks they should
be in Debian should rather advocate them (and hence be ruled out from
AMing), but placing this as a requirement here will rule out a good
number of otherwise valid assignments. Also, this was observed for
returning emerituses which do not need advocates, and the restriction
makes even less sense.