Enhance RT ticket template to use Front Desk for DMs
Those of us who aren't DAM would like to be able to use the RT
templates available from the site, such as for requesting the
addition of new DMs. Alter the template to talk about Front Desk
for DMs, and DAM for everything else.

(Potential future change: Also Front Desk for guest accounts.)